+++ src/crayon/.npmignore
