+++ src/crayon/__tests__/mocks/document.ts
