+++ src/crayon/__tests__/mocks/window.ts
