+++ src/crayon/src/index.ts
@@ -4,10 +4,10 @@ export * from './platform/history'
 export * from './platform/locator'
 export * from './platform/route-map'
 export * from './platform/shared-state'
-export * from './api'
+export * from './sdk'
 
 
-import { create, group } from './api'
+import { create, group } from './sdk'
 import { Router as OGRouter, Group as OGGroup } from './platform/router'
 
 export const crayon = {
