+++ src/crayon/src/platform/mount/has-animation.ts
