+++ src/crayon/src/platform/mount/mount-static.ts
