+++ src/crayon/src/platform/router/response.ts
